rename authentication.context to UserAuthenticationContext
