+++ src/components/protectedRoutes/ProtectedRoutesUser.jsx
@@ -1,6 +1,6 @@
 import { useContext } from "react";
 import { Navigate, Outlet } from "react-router-dom";
-import { AuthenticationContext } from "../services/authentication/authentication.context";
+import { AuthenticationContext } from "../services/authentication/UserAuthenticationContext";
 
 const ProtectedRoutesUser = () => {
